masterlibs: AP_HAL: Util: refactored get_system_clock_utc funct.

+++ masterlibs/libraries/AP_HAL/Util.h
@@ -54,7 +54,7 @@ public:
     /*
       get system time in UTC hours, minutes, seconds and milliseconds
      */
-    void get_system_clock_utc(int32_t &hour, int32_t &min, int32_t &sec, int32_t &ms) const;
+    void get_system_clock_utc(uint16_t &hour, uint8_t &min, uint8_t &sec, uint16_t &ms) const;
 
     uint32_t get_time_utc(int32_t hour, int32_t min, int32_t sec, int32_t ms) const;
